+++ src/components/Layout/Header/ThemeSelector.tsx
@@ -6,25 +6,29 @@ import { RiSunLine, RiMoonLine, RiPaintBrushLine, RiCheckDoubleLine } from 'reac
 interface IThemeSelectorProps {
   visible: boolean;
   newLabel?: string;
+  onChange?: (isDark: boolean, theme: string) => void;
 }
 
-const ThemeSelector: FC<IThemeSelectorProps> = ({ visible, newLabel }) => {
+const ThemeSelector: FC<IThemeSelectorProps> = ({ visible, newLabel, onChange }) => {
   const { theme, setTheme } = useThemeContext();
   if (!visible) return null;
 
   const onChangeTheme = (isDark: boolean, theme: string) => {
     setTheme(theme, isDark);
+    onChange?.(isDark, theme);
   };
 
   return (
     <div className="relative mx-auto flex max-w-screen-xl justify-end px-4 pt-3 md:px-8 lg:px-10">
       <div className="flex flex-wrap items-center justify-center gap-2 md:justify-start md:space-x-4 ">
-        <span className="inline-block rounded-full border border-primary-400 px-2 py-2 text-xs font-medium text-primary-800">
+        {/*<span className="inline-block rounded-full border border-primary-400 px-2 py-2 text-xs font-medium text-primary-800">
           {newLabel} <RiPaintBrushLine className="ml-1 inline-block" />
-        </span>
+        </span>*/}
         <div className="flex flex-wrap items-center justify-center gap-3">
           <ThemeOption isDark={true} theme="main-dark" onClick={onChangeTheme} currentTheme={theme} />
+          <ThemeOption isDark={true} theme="orange-dark" onClick={onChangeTheme} currentTheme={theme} />
           <ThemeOption isDark={false} theme="main-light" onClick={onChangeTheme} currentTheme={theme} />
+          <ThemeOption isDark={false} theme="orange-light" onClick={onChangeTheme} currentTheme={theme} />
         </div>
       </div>
     </div>
@@ -49,7 +53,7 @@ const ThemeOption: FC<IThemeOptionProps> = ({ isDark, theme, onClick, currentThe
         )}
       </div>
       {currentTheme === theme && (
-        <span className="absolute top-0 right-0 -mt-1 -mr-1 inline-flex items-center justify-center rounded-full bg-primary-500 px-1 py-1 text-xs font-medium leading-none text-primary-500">
+        <span className="absolute right-0 top-0 -mr-1 -mt-1 inline-flex items-center justify-center rounded-full bg-primary-500 px-1 py-1 text-xs font-medium leading-none text-primary-500">
           <span className="sr-only">Selected</span>
           <RiCheckDoubleLine className="h-2 w-2 text-primaryText-500" />
         </span>